+++ FishInABarrel/Patches/TerminalPatch.cs
@@ -2,12 +2,15 @@
 
 namespace FishInABarrel.Patches
 {
+	/// <summary>
+	/// Infinite credits
+	/// </summary>
 	[HarmonyPatch(typeof(Terminal))]
 	internal class TerminalPatch
 	{
 		[HarmonyPatch("BeginUsingTerminal")]
 		[HarmonyPostfix]
-		private static void onBeginUsingTerminal(ref int ___groupCredits)
+		private static void PostBeginUsingTerminal(ref int ___groupCredits)
 		{
 			___groupCredits = 9999;
 		}