+++ Store Bundle/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 Here's what I've been up to...
 
+## v1.0.4
+- Inverse teleporter does not remove player inventory
+- Keys are not removed from inventory after use
+- Infinite boombox battery
+
 ## v1.0.3
 - Inverse teleporter is on the ship at the beginning of the game
 - Remove beehives